+++ CHANGELOG.md
@@ -7,6 +7,10 @@ _2025-2-11_
 - New beacon type added for drop beacon
 - New beacon type and sub type support for performance beacon
 
+_2025-2-17_
+
+- Added RateLimits feature to customize beacon transmission frequency.
+
 ## Version 6.1.0
 
 _2024-11-27_